// shopapp-angular/src/app/models/product.ts
import { ProductImage } from "./product.image";
export interface Product {
  id: number;
  name: string;
  price: number;
  thumbnail: string;
  description: string;
  category_id: number;
  url: string;
  product_images: ProductImage[];
  quantityInStock?: number; // Số lượng trong kho, có thể không có
  soldQuantity?: number; // Số lượng đã bán, có thể không có
}

  
  